restructuring code
restructured the code for checking for an enemy. Now the data has a list of targets instead of just a single variable. Now i use a for each loop to check through that array and if the array has the name of the target i'm looking at it works. makes it so a single unit can attack more than one type of target